requirements.txt
fastapi
uvicorn
langchain-google-genai
python-dotenv
nest_asyncio
requests
yfinance
